requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.2
allora_sdk==0.2.3
annotated-types==0.7.0
anyio==4.9.0
asyncio_rpc==0.3.2
attrs==25.3.0
backoff==2.2.1
base58==2.1.1
bcl==2.3.1
bcrypt==4.3.0
bitarray==3.4.3
boto3==1.39.3
boto3-stubs==1.39.3
botocore==1.39.3
botocore-stubs==1.38.46
bytecode==0.16.2
cachetools==5.5.0
cdp-sdk==1.19.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.2.1
coinbase-agentkit==0.6.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.6.2
cryptography==45.0.5
cytoolz==1.0.1
datasets==2.21.0
ddtrace==2.21.0
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
egcd==2.0.2
email_validator==2.2.0
envier==0.6.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.111.1
fastapi-cli==0.0.7
filelock==3.16.1
fire==0.7.0
fireworks-ai==0.15.15
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
h11==0.16.0
hexbytes==1.3.1
hf-xet==1.1.5
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
httpx-ws==0.7.2
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonalias==0.1.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lagrange==3.0.1
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
legacy-cgi==2.6.3
litellm==1.74.0
loguru==0.7.3
lz4==4.4.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
mypy-boto3==1.39.0
mypy-boto3-s3==1.39.2
nearai==0.1.19
nest-asyncio==1.6.0
nilql==0.0.0a13
numpy==2.3.1
oauthlib==3.3.1
openai==1.93.0
opentelemetry-api==1.34.1
orjson==3.10.18
packaging==24.2
pailliers==0.2.0
pandas==2.3.0
pandas-stubs==2.3.0.250703
paramiko==3.5.1
parsimonious==0.10.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.3.2
protobuf==6.31.1
psutil==5.9.8
py-near==1.1.57
py_near_primitives==0.2.4
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyproject-api==1.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
rabinmiller==0.1.0
redis==4.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rlp==4.1.0
rpds-py==0.26.0
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
solana==0.36.7
solders==0.26.0
SQLAlchemy==2.0.41
starlette==0.37.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toolz==1.0.0
tox==4.23.2
tqdm==4.67.1
tweepy==4.16.0
typer==0.12.5
types-awscrt==0.27.4
types-cffi==1.17.0.20250523
types-colorama==0.4.15.20240311
types-docutils==0.21.0.20250604
types-psutil==6.1.0.20241221
types-Pygments==2.19.0.20250516
types-pyOpenSSL==24.1.0.20240722
types-python-dateutil==2.9.0.20250516
types-pytz==2025.2.0.20250516
types-PyYAML==6.0.12.20250516
types-redis==4.6.0.20241004
types-regex==2024.11.6.20250403
types-requests==2.32.4.20250611
types-s3transfer==0.13.0
types-setuptools==71.1.0.20240818
types-ujson==5.10.0.20250326
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.28.1
watchfiles==1.1.0
wcwidth==0.2.13
web3==7.10.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
